Move createStore to internal
