+++ src/create-context.ts
@@ -1,4 +1,5 @@
-import { Component, ContextStore, Element } from './types';
+import { createStore } from './internal/create-store';
+import { Component, Element } from './types';
 
 export interface ProviderProps<C = {}> {
   context: C;
@@ -13,37 +14,6 @@ export interface ConsumerProps<C = {}> {
 
 export interface ConsumerState<C> {
   context: C;
-}
-
-function createStore<C = {}>(initialContext: C): ContextStore<C> {
-  let context = initialContext;
-
-  const store: ContextStore<C> = {
-    subscribers: [],
-    subscribe: callback => {
-      store.subscribers.push(callback);
-
-      return () => {
-        const index = store.subscribers.indexOf(callback);
-
-        if (index >= 0) {
-          store.subscribers.splice(index, 1);
-        }
-      };
-    },
-    setContext: (newContext: C) => {
-      context = newContext;
-
-      store.subscribers.forEach(subscriber => {
-        subscriber(context);
-      });
-    },
-    getContext: () => {
-      return context;
-    },
-  };
-
-  return store;
 }
 
 export function createContext<C = {}>(initialContext: C) {
